Changed long double on last advanced of functions and condition on dates

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -5,14 +5,14 @@
 */
 int main(void)
 {
-	unsigned long int n1 = 1, n2 = 2, r=0;
+	long double n1 = 1, n2 = 2, r = 0;
 	int i;
 
 	printf("1, 2, ");
 	for (i = 0; i < 96; i++)
 	{
 		r = n1 + n2;
-		printf("%lu", r);
+		printf("%.0Lf", r);
 		n1 = n2;
 		n2 = r;
 		if (i != 95)
